requirements.txt
pandas
python-dotenv
os-sys

openai
langchain
langchain-openai
langchain-google-genai
langchain-chroma
langchain-community
langsmith

jq
pyjq

streamlit
faiss-cpu
beautifulsoup4
tiktoken
pypdf
PyPDF2

DateTime
typing
requests